DSWNIFI 1.3 update:
1) wifi_arm9.c :
- protect some structs in case DSWIFI deinit code calls them even if
those
are not inited (fixes data aborts)
- update gdbstub, now runs as part of dswnifi since gdb server can´t
operate
while local/udp nifi is working

2) merge GDB debugging into dswnifi selector
3) some cleanup

These changes cause the following to happen:
a) Allows DS to re-connect from AP if disconnected. (The AP still holds
the original DS IP it acquired)
DSWIFI is re-inited properly.

b) due to a):
Allows to change from WIFI AP mode to adhoc, or adhoc to WIFI AP, or
disable completely DSWIFI on the fly.